# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Palace HTTP API",
    "description": "Owner-scoped Palace API. Schema describes wire structure, not authorization or transactional guarantees. Body limits use injected ImportLimits (default raw history 8 MiB); HTTP limit is bytes * 6 + 64 KiB. UTF-8 byte limits, nonblank strings and canonical cursor upper bounds are enforced by the server. Native extractor errors may be text/plain. No deployment credentials or environment metadata are included.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
    "/api/conversations/{id}": {
      "get": {
        "tags": [
          "business"
        ],
        "summary": "Returns a stable tree and controlled external link only for the authenticated owner.",
        "operationId": "conversation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Owner-scoped UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "Refreshed production session; absent in fixed-user mode"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input; JSON domain error or native extractor text as declared",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InputErrorResponse"
                }
              },
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable in owner scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Identity, source session or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal persistence or response failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Identity provider temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "path_management"
        ],
        "summary": "Deletes the card and all branches atomically inside the authenticated owner's scope.",
        "operationId": "delete_conversation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Owner-scoped UUID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Origin",
            "in": "header",
            "description": "Exactly one header matching the configured external origin; null, duplicates and omissions are rejected",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "headers": {
              "Cache-Control": {
                "schema": {
                  "type": "string"
                },
                "description": "no-store"
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "Refreshed production session; absent in fixed-user mode"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deleted"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input; JSON domain error or native extractor text as declared",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Origin rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Resource unavailable in owner scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Identity, source session or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal persistence or response failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Identity provider temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/auth/login": {
      "get": {
        "tags": [
          "handlers"
        ],
        "summary": "Starts a persisted, browser-bound Authorization Code Flow.",
        "operationId": "login",
        "responses": {
          "303": {
            "description": "Browser redirect; empty body",
            "headers": {
              "Location": {
                "schema": {
                  "type": "string"
                },
                "description": "Identity provider authorization URL"
              },
              "Set-Cookie": {
                "schema": {
                  "type": "string"
                },
                "description": "Secure HttpOnly browser credential"
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Identity, source session or idempotency conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal persistence or response failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Identity provider temporarily unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
      "Conversation": {
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "source"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid"
          },
          "source": {
            "$ref": "#/components/schemas/Source"
          }
        }
      },
      "ConversationDetail": {
        "type": "object",
        "required": [
          "conversation",
          "messages",
          "paths"
        ],
        "properties": {
          "conversation": {
            "$ref": "#/components/schemas/Conversation"
          },
          "messages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          },
          "paths": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ConversationPath"
            }
          }
        }
      },
...
      "ConversationPath": {
        "type": "object",
        "required": [
          "id",
          "title",
          "session_id",
          "head_message_id",
          "occurred_at",
          "created_at",
          "updated_at",
          "message_count",
          "original_link"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix epoch milliseconds assigned by the database."
          },
          "head_message_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "message_count": {
            "type": "integer",
            "format": "int64"
          },
          "occurred_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix epoch milliseconds selected by the user."
          },
          "original_link": {
            "type": "string",
            "description": "A controlled external link, never fetched by the server."
          },
          "session_id": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "description": "The title of this source session; other paths of the tree have their own."
          },
          "updated_at": {
            "type": "integer",
            "format": "int64",
            "description": "Unix epoch milliseconds assigned by the database."
          }
        }
      },
...
      "Deleted": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "ErrorCode": {
        "type": "string",
        "enum": [
          "authentication_required",
          "origin_rejected",
          "identity_or_request_conflict",
          "session_already_exists",
          "not_found",
          "identity_unavailable",
          "persistence_failed"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error",
          "login"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "login": {
            "type": "string",
            "description": "/auth/login for authentication_required; the empty string for other errors. Always present."
          }
        }
      },
...
      "InputErrorKind": {
        "type": "string",
        "enum": [
          "syntax",
          "field",
          "limit"
        ]
      },
      "InputErrorResponse": {
        "type": "object",
        "required": [
          "kind",
          "path",
          "message"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/InputErrorKind"
          },
          "message": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
...
      "Message": {
        "type": "object",
        "required": [
          "id",
          "owner_id",
          "conversation_id",
          "parent_message_id",
          "role",
          "content",
          "created_order"
        ],
        "properties": {
          "content": {
            "type": "string"
          },
          "conversation_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_order": {
            "type": "integer",
            "format": "int64",
            "description": "Ordering integer, not a timestamp; i64 values are not generally lossless JS numbers."
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "owner_id": {
            "type": "string",
            "format": "uuid"
          },
          "parent_message_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Always present; root messages carry null."
          },
          "role": {
            "$ref": "#/components/schemas/Role"
          }
        }
      },
...
      "Role": {
        "type": "string",
        "enum": [
          "user",
          "assistant"
        ]
      },
      "Source": {
        "type": "string",
        "enum": [
          "chatgpt",
          "gemini",
          "grok"
        ]
      },
...
    },
    "securitySchemes": {
      "login_flow": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-palace-login",
        "description": "Browser-bound one-time login flow; validated together with query state before exchanging code."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-palace-session",
        "description": "Production browser session; Secure, HttpOnly, SameSite=Lax, Path=/. Business handlers receive a server-resolved owner. Local fixed-user mode is a development-only boundary."
      }
    }
  }
}